requirements.txt
APScheduler==3.10.4
autobahn==24.4.2
beautifulsoup4==4.12.3
colorama==0.4.6
cssmin==0.2.0
Flask==3.0.3
Flask-Assets2==2.1
irc==20.5.0
isodate==0.6.1
Jinja2==3.1.4
jsmin==3.0.1
Markdown==3.7
MarkupSafe==2.1.5
marshmallow-dataclass==8.7.1
marshmallow==3.22.0
psycopg2==2.9.10
pylast==5.3.0
pyOpenSSL==24.2.1 # for twisted.internet.ssl
pytz==2024.2
rapidfuzz==3.10.0
redis==5.0.8
regex==2024.9.11
requests==2.32.3
retry==0.9.2
riotwatcher==3.3.0
service-identity==24.1.0 # for twisted SSL as well
SQLAlchemy==2.0.38
SQLAlchemy-Utc==0.14.0
tempora==5.7.0
Twisted==24.7.0
Unidecode==1.3.8
uWSGI==2.0.27
urlextract==1.9.0
Flask-WTF==1.2.1
pytimeparse==1.1.8
legacy-cgi==2.6.2
grpcio==1.70.0
grpcio_tools==1.70.0
grpcio-reflection==1.70.0
